+++ src/plonkish/types.rs
@@ -1,0 +1,6 @@
+use ark_ff::Field;
+
+#[derive(Clone)]
+pub struct Constraint<F: Field> {
+  pub points: Vec<F>,
+}
